pyproject.toml
[tool.poetry]
name = "langchain-chatbot"
version = "0.1.0"
description = ""
authors = ["Ali Noroozi"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
langchain-core = "^0.1.34"
langchain-openai = "^0.1.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
